Move release link from readme to branch selector
