+++ source/content.js
@@ -32,6 +32,7 @@ import addProfileHotkey from './features/add-profile-hotkey';
 import addYoursMenuItem from './features/add-yours-menu-item';
 import addToggleFilesButton from './features/add-toggle-files-button';
 import addReadmeButtons from './features/add-readme-buttons';
+import addBranchButtons from './features/add-branch-buttons';
 import addDeleteForkLink from './features/add-delete-fork-link';
 import linkifyIssuesInTitles from './features/linkify-issues-in-titles';
 import addPatchDiffLinks from './features/add-patch-diff-links';
@@ -167,6 +168,7 @@ function ajaxedPagesHandler() {
 
 	if (pageDetect.isRepo()) {
 		enableFeature(addReadmeButtons);
+		enableFeature(addBranchButtons);
 		enableFeature(addDiffViewWithoutWhitespaceOption);
 		enableFeature(removeDiffSigns);
 		enableFeature(addCILink);
